feat(backend): parseをblockごとではなくblockの配列ごとに行うように変更
